+++ tests/test_potentials.py
@@ -22,7 +22,7 @@ def test_harmonic():
     atoms = Atoms(atom, pos, a=a, ecut=ecut)
     atoms.s = s
     atoms.f = f
-    E = RSCF(atoms, pot=pot, guess="random", etol=1e-6).run()
+    E = RSCF(atoms, pot=pot, etol=1e-6).run()
     # We have to get close to the Etot reference value of 43.337 Eh (for different parameters)
     assert_allclose(E, 43.10344)
 
@@ -38,10 +38,50 @@ def test_coulomb():
 
     atoms = Atoms(atom, pos, a=a, ecut=ecut)
     atoms.s = s
-    E = RSCF(atoms, pot=pot, guess="random", etol=1e-6).run()
+    E = RSCF(atoms, pot=pot, etol=1e-6).run()
     # In the limit we should come close to the NIST Etot value of -0.445671 Eh
     # See https://nist.gov/pml/atomic-reference-data-electronic-structure-calculations/atomic-reference-data-electronic-7-0
     assert_allclose(E, -0.4393708)
+
+
+def test_coulomb_species():
+    """Compare total energies for the ionic potential with different species."""
+    atom = "LiH"
+    pos = ((0, 0, 0), (1, 0, 0))
+    a = 10
+    ecut = 5
+    Z = [3, 1]
+
+    atoms = Atoms(atom, pos, a=a, ecut=ecut)
+    atoms.Z = Z
+    E_GTH = RSCF(atoms, pot="GTH", etol=1e-6).run()
+    E_C = RSCF(atoms, pot="Coulomb", etol=1e-6).run()
+    # The all-electron GTH and Coulomb energies should be somewhat similar
+    assert_allclose(E_GTH, E_C, atol=1e-2)
+
+
+def test_coulomb_lr():
+    """Compare total energies for the ionic potential with different species."""
+    atom = "H"
+    pos = (0, 0, 0)
+    a = 10
+    ecut = 5
+
+    atoms = Atoms(atom, pos, a=a, ecut=ecut)
+    E_C = RSCF(atoms, pot="Coulomb", etol=1e-6).run()
+    scf = RSCF(atoms, pot="LR", etol=1e-6)
+    scf.pot_params = {"alpha": 20}
+    E_LR = scf.run()
+
+    # For smaller alpha the long-range energy should be smaller than the normal Coulomb energy
+    assert_allclose(E_C, E_LR, atol=1e-3)
+    assert E_C < E_LR
+
+    scf = RSCF(atoms, pot="LR", etol=1e-6)
+    scf.pot_params = {"alpha": 10000}
+    E_LR = scf.run()
+    # For large alpha the potentials should be identical
+    assert_allclose(E_C, E_LR)
 
 
 def test_ge():
@@ -57,7 +97,7 @@ def test_ge():
     atoms = Atoms(atom, pos, a=a, ecut=ecut)
     atoms.s = s
     atoms.f = f
-    scf = RSCF(atoms, pot=pot, guess="random", etol=1e-6)
+    scf = RSCF(atoms, pot=pot, etol=1e-6)
     scf.run()
     eps = get_epsilon(scf, scf.W)[0]
     # In the limit we should come close to the NIST 4s-4p value of -0.276641 Eh